working read not working - create, update, delete
some fixes, still not working

+++ src/main/java/com/in3des/springlesson/controllers/PeopleController.java
@@ -1,7 +1,6 @@
 package com.in3des.springlesson.controllers;
 
 import com.in3des.springlesson.dao.PersonDAO;
-import com.in3des.springlesson.dao.PersonDAOImpl;
 import com.in3des.springlesson.entity.Person;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -29,7 +28,7 @@ public class PeopleController {
     }
 
     @GetMapping("/{id}")
-    public String show(@PathVariable("id") int id, Model model) {
+    public String show(@PathVariable("id") Long id, Model model) {
         model.addAttribute("person", personDAO.show(id));
         return "people/show";
     }
@@ -52,14 +51,14 @@ public class PeopleController {
     }
 
     @GetMapping("/{id}/edit")
-    public String edit(@PathVariable("id") int id, Model model) {
+    public String edit(@PathVariable("id") Long id, Model model) {
         model.addAttribute("person", personDAO.show(id));
         return "people/edit";
     }
 
     @PatchMapping("/{id}")
     public String update(@ModelAttribute("person") @Valid Person person, BindingResult bindingResult,
-                         @PathVariable("id") int id) {
+                         @PathVariable("id") Long id) {
 
         if (bindingResult.hasErrors()) {
             return "people/edit";
@@ -70,7 +69,7 @@ public class PeopleController {
     }
 
     @DeleteMapping("/{id}")
-    public String delete(@PathVariable("id") int id) {
+    public String delete(@PathVariable("id") Long id) {
         personDAO.delete(id);
         return "redirect:/people";
     }
