+++ src/main/java/com/in3des/springlesson/dao/PersonDAO.java
@@ -3,9 +3,6 @@ package com.in3des.springlesson.dao;
 import com.in3des.springlesson.entity.Person;
 import org.springframework.stereotype.Repository;
 
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import javax.transaction.Transactional;
 import java.util.List;
 
 //@Component
@@ -15,13 +12,13 @@ public interface PersonDAO {
 
     public List<Person> index();
 
-    public Person show(final int id);
+    public Person show(final Long id);
 
     public void save(Person person);
 
-    public void update(Person updatedPerson, int id);
+    public void update(Person updatedPerson, Long id);
 
-    public void delete(int id);
+    public void delete(Long id);
 
 
 }
