+++ src/main/java/com/in3des/springlesson/service/PeopleService.java
@@ -1,8 +1,6 @@
 package com.in3des.springlesson.service;
 
-import com.in3des.springlesson.dao.PersonDAOImpl;
 import com.in3des.springlesson.entity.Person;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 
@@ -10,11 +8,11 @@ public interface PeopleService {
 
     public List<Person> index();
 
-    public Person show(final int id);
+    public Person show(final Long id);
 
     public void save(Person person);
 
-    public void update(Person updatedPerson, int id);
+    public void update(Person updatedPerson, Long id);
 
-    public void delete(int id);
+    public void delete(Long id);
 }
